LayoutTests/fast/workers/worker-inherits-csp-blocks-xhr-expected.txt

CONSOLE MESSAGE: Refused to connect to 'non-existent-file' because it violates the following Content Security Policy directive: "connect-src 'none'".

This tests that the Content Security Policy (CSP) of the owner document (this page) blocks a file-URL Web Worker from making an XHR request because the parent's CSP contains "connect-src 'none'"

PASS threw exception Error: SecurityError: DOM Exception 18.
